common/requirements.txt
# The following libraries are optional and need to be install manually.
# django==1.6.11
# MySQL-Python
# M2Crypto
# pylibmc
# pymssql
# ssdb

asgiref==3.3.4
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
confluent-kafka==1.7.0
croniter==1.0.13
cryptography==3.4.7
Cython==0.29.23
Django==3.0.4
django-mssql-backend==2.8.1
dnspython==1.16.0
eventlet==0.31.0
firebase-admin==2.1.1
Flask==2.0.1
futures==3.1.1
gevent==21.1.2
gevent-fastcgi==1.1.0.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
greenlet==1.1.0
gspread==3.7.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
IP2Location==8.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
kafka-python==2.0.2
kazoo==2.8.0
lxml==4.6.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysqlclient==2.0.3
ndg-httpsclient==0.5.1
oauth2client==4.1.3
oauthlib==3.1.1
paramiko==2.7.2
phonenumbers==8.12.24
Pillow==8.2.0
pip==21.1.2
prometheus-client==0.7.1
protobuf==2.6.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pycurl==7.43.0.6
pygeoip==0.3.2
pymemcache==3.4.4
PyMySQL==1.0.2
PyNaCl==1.4.0
pyodbc==4.0.30
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pysftp==0.2.8
python-dateutil==2.8.1
python-memcached==1.59
pytz==2021.1
qrcode==6.1
raven==6.10.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
simplejson==3.17.2
six==1.16.0
SQLAlchemy==1.4.18
sqlparse==0.4.1
tornado==4.1
urllib3==1.26.5
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
